General update and desync fixes

+++ Assets/_Project/Scripts/CollisionBoxScripts/CollisionBoxVisualizerScripts/CollisionVisualizer.cs
@@ -9,7 +9,7 @@ public class CollisionVisualizer : MonoBehaviour
     {
         if (TrainingSettings.ShowHitboxes)
         {
-            Vector2 fixedPosition = new Vector2Int(Mathf.FloorToInt(collider.position.x * 1) / 1, Mathf.FloorToInt(collider.position.y * 1) / 1);
+            Vector2 fixedPosition = new Vector2Int(Mathf.FloorToInt((float)collider.position.x * 1) / 1, Mathf.FloorToInt((float)collider.position.y * 1) / 1);
             transform.position = fixedPosition;
             _spriteRenderer.size = new Vector2((float)collider.size.x, (float)collider.size.y);
             _spriteRenderer.enabled = collider.active;
